refactored codebase pt.1
