+++ client/.eslintrc.json
@@ -19,5 +19,6 @@
         "react"
     ],
     "rules": {
+        "react/prop-types": 0
     }
 }
