+++ client/package.json
@@ -47,5 +47,6 @@
     "webpack": "^5.65.0",
     "webpack-cli": "^4.9.1",
     "webpack-dev-server": "^4.6.0"
-  }
+  },
+  "proxy": "http://localhost:4000"
 }
