+++ client/src/components/ActionBtn.js
@@ -1,11 +1,13 @@
+import React from 'react';
+
 const ActionBtn = (props) => {
-  return (
-    <>
-      <button className='action-btn' {...props}>
-        {props.children}
-      </button>
-    </>
-  );
+    return (
+        <>
+            <button className='action-btn' {...props}>
+                {props.children}
+            </button>
+        </>
+    );
 };
 
 export default ActionBtn;
