+++ client/src/components/Filter.js
@@ -1,3 +1,4 @@
+import React from 'react';
 import FilterButton from './FilterButton';
 
 const Filter = (props) => {
