+++ client/src/components/FilterButton.js
@@ -1,11 +1,13 @@
+import React from 'react';
+
 const FilterButton = (props) => {
-  return (
-    <>
-      <button className='filter-btn' key={props.id} {...props}>
-        {props.children}
-      </button>
-    </>
-  );
+    return (
+        <>
+            <button className='filter-btn' key={props.id} {...props}>
+                {props.children}
+            </button>
+        </>
+    );
 };
 
 export default FilterButton;
