+++ client/src/components/FooterLink.js
@@ -1,3 +1,4 @@
+import React from 'react';
 import { Link } from 'react-router-dom';
 
 const MenuLink = ({ id, url, text }) => {
