+++ client/src/components/Form.js
@@ -1,9 +1,11 @@
+import React from 'react';
+
 const Form = (props) => {
-  return (
-    <>
-      <form {...props}>{props.children}</form>
-    </>
-  );
+    return (
+        <>
+            <form {...props}>{props.children}</form>
+        </>
+    );
 };
 
 export default Form;
