+++ client/src/components/Loading.js
@@ -1,11 +1,13 @@
+import React from 'react';
+
 const Loading = () => {
-  return (
-    <>
-      <div className='loading'>
-        <div className='loader'></div>
-      </div>
-    </>
-  );
+    return (
+        <>
+            <div className='loading'>
+                <div className='loader'></div>
+            </div>
+        </>
+    );
 };
 
 export default Loading;
