+++ client/src/components/Loading2.js
@@ -1,14 +1,17 @@
+import React from 'react';
 import { useGlobalContext } from '../context/GlobalContext';
 
 const Loading2 = () => {
-  const { loading } = useGlobalContext();
-  return (
-    <>
-      <div className={`${loading ? 'load-wrapper open' : 'load-wrapper'}`}>
-        <div className='loader'></div>
-      </div>
-    </>
-  );
+    const { loading } = useGlobalContext();
+    return (
+        <>
+            <div
+                className={`${loading ? 'load-wrapper open' : 'load-wrapper'}`}
+            >
+                <div className='loader'></div>
+            </div>
+        </>
+    );
 };
 
 export default Loading2;
