+++ client/src/components/MenuLink.js
@@ -1,3 +1,4 @@
+import React from 'react';
 import { NavLink } from 'react-router-dom';
 import { useGlobalContext } from '../context/GlobalContext';
