+++ client/src/components/Message.js
@@ -1,3 +1,4 @@
+import React from 'react';
 import { FaUserCircle } from 'react-icons/fa';
 import { useDateFormatter } from '../hooks/useDateFormatter';
