+++ client/src/components/ReturnTo.js
@@ -1,9 +1,10 @@
+import React from 'react';
 import { useNavigate } from 'react-router-dom';
 import FilterButton from './FilterButton';
 
 const ReturnTo = () => {
     const history = useNavigate();
-    const prevPath = () => history.goBack();
+    const prevPath = () => history(-1);
     return (
         <>
             <FilterButton onClick={prevPath}>zurück</FilterButton>
