+++ client/src/components/SearchBar.js
@@ -1,21 +1,19 @@
-import { useMarketplaceContext } from '../context/MarketplaceContext';
+import React from 'react';
 
-const SearchBar = () => {
-  const { search, handleSearch } = useMarketplaceContext();
-
-  return (
-    <>
-      <section className='search-bar'>
-        <input
-          type='text'
-          className='search-form'
-          value={search}
-          onChange={handleSearch}
-          placeholder='Nach Titel oder Autor*in suchen...'
-        ></input>
-      </section>
-    </>
-  );
+const SearchBar = ({ search, handleSearch }) => {
+    return (
+        <>
+            <section className='search-bar'>
+                <input
+                    type='text'
+                    className='search-form'
+                    value={search}
+                    onChange={handleSearch}
+                    placeholder='Nach Titel oder Autor*in suchen...'
+                ></input>
+            </section>
+        </>
+    );
 };
 
 export default SearchBar;
