+++ client/src/components/SigninBtn.js
@@ -1,11 +1,13 @@
+import React from 'react';
+
 const SigninBtn = (props) => {
-  return (
-    <>
-      <button className='signin-btn' {...props}>
-        {props.children}
-      </button>
-    </>
-  );
+    return (
+        <>
+            <button className='signin-btn' {...props}>
+                {props.children}
+            </button>
+        </>
+    );
 };
 
 export default SigninBtn;
