+++ client/src/components/TextAreaInput.js
@@ -1,14 +1,16 @@
+import React from 'react';
+
 const TextAreaInput = (props) => {
-  return (
-    <>
-      <div className='form-control'>
-        <label htmlFor={props.htmlFor} name={props.name}>
-          {props.htmlFor}
-        </label>
-        <textarea {...props} />
-      </div>
-    </>
-  );
+    return (
+        <>
+            <div className='form-control'>
+                <label htmlFor={props.htmlFor} name={props.name}>
+                    {props.htmlFor}
+                </label>
+                <textarea {...props} />
+            </div>
+        </>
+    );
 };
 
 export default TextAreaInput;
