+++ client/src/pages/MyBooks.js
@@ -1,14 +1,47 @@
+import React, { useState, useCallback, useEffect } from 'react';
 import { useGlobalContext } from '../context/GlobalContext';
-import { useMyBooksContext } from '../context/MyBooksContext';
 import UserDashboard from '../components/UserDashboard';
 import Loading from '../components/Loading';
 import Alert from '../components/Alert';
 import { motion } from 'framer-motion';
 import Shelf from '../components/Shelf';
+import { API_BOOKSBYUSER } from '../config/config';
 
 const MyBooks = () => {
     const { alert, closeSubmenu, loading } = useGlobalContext();
-    const { myBooks } = useMyBooksContext();
+    const [myBooks, setMyBooks] = useState([]);
+    const { setLoading, userId, jwt } = useGlobalContext();
+
+    // GET Bücher des Users
+    const fetchMyBooks = useCallback(
+        async (api, id, token) => {
+            setLoading(true);
+            try {
+                const res = await fetch(`${api}${id}`, {
+                    headers: {
+                        Authorization: `Bearer ${token}`,
+                        'content-type': 'application/json'
+                    }
+                });
+                if (res.ok) {
+                    const myBookList = await res.json();
+                    setMyBooks(myBookList);
+                } else {
+                    throw new Error(`could not get books of user ${id}`);
+                }
+            } catch (err) {
+                console.log(err);
+            } finally {
+                setLoading(false);
+            }
+        },
+        [setLoading]
+    );
+
+    // hole Bücher des Users
+    useEffect(() => {
+        fetchMyBooks(API_BOOKSBYUSER, userId, jwt);
+    }, [fetchMyBooks, userId, jwt]);
 
     return (
         <>
@@ -23,7 +56,7 @@ const MyBooks = () => {
                     onClick={closeSubmenu}
                 >
                     <UserDashboard />
-                    <Shelf>{myBooks}</Shelf>
+                    <Shelf element={myBooks} />
                     {alert.display && <Alert />}
                 </motion.main>
             )}
